Use new directory structure for components and containers
