+++ webapp/components/sidebar.jsx
@@ -4,7 +4,7 @@
 import $ from 'jquery';
 import ReactDOM from 'react-dom';
 import NewChannelFlow from './new_channel_flow.jsx';
-import MoreDirectChannels from 'containers/more_direct_channels';
+import MoreDirectChannels from 'components/more_direct_channels';
 import MoreChannels from 'components/more_channels.jsx';
 import SidebarHeader from './sidebar_header.jsx';
 import UnreadChannelIndicator from './unread_channel_indicator.jsx';
