--- webapp/containers/user_settings_general.js
@@ -1,24 +1,0 @@
-// Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
-// See License.txt for license information.
-
-import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
-import {getMe} from 'mattermost-redux/actions/users';
-
-import UserSettingsGeneralTab from 'components/user_settings/user_settings_general.jsx';
-
-function mapStateToProps(state, ownProps) {
-    return {
-        ...ownProps
-    };
-}
-
-function mapDispatchToProps(dispatch) {
-    return {
-        actions: bindActionCreators({
-            getMe
-        }, dispatch)
-    };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(UserSettingsGeneralTab);
